moved files from "update IAM policy samples to use GAPIC library" PR (#9319)
* moved all files from "update IAM policy samples to use GAPIC library" PR to snippets module

* merge with main, create TestPermissionsTest

* remove unused test case

* added license header

* fix lint

* fix lint
