+++ iam/api-client/pom.xml
@@ -109,12 +109,6 @@
       <artifactId>google-iam-admin</artifactId>
       <scope>compile</scope>
     </dependency>
-    <dependency>
-      <groupId>com.google.cloud</groupId>
-      <artifactId>google-cloud-resourcemanager</artifactId>
-      <scope>compile</scope>
-    </dependency>
-
   </dependencies>
 
   <build>
@@ -124,7 +118,7 @@
         <artifactId>exec-maven-plugin</artifactId>
         <version>3.1.1</version>
         <configuration>
-          <mainClass>iam.snippets.Quickstart</mainClass>
+          <mainClass>iam.snippets.CreateServiceAccount</mainClass>
         </configuration>
       </plugin>
     </plugins>
