+++ iam/snippets/src/main/java/GetServiceAccountPolicy.java
@@ -13,34 +13,44 @@
  * limitations under the License.
  */
 
-package iam.snippets;
+// [START iam_service_account_get_policy]
 
-// [START iam_get_policy]
-import com.google.cloud.resourcemanager.v3.ProjectsClient;
-import com.google.iam.admin.v1.ProjectName;
+import com.google.cloud.iam.admin.v1.IAMClient;
+import com.google.iam.admin.v1.ServiceAccountName;
 import com.google.iam.v1.GetIamPolicyRequest;
 import com.google.iam.v1.Policy;
 import java.io.IOException;
 
-public class GetProjectPolicy {
+public class GetServiceAccountPolicy {
   public static void main(String[] args) throws IOException {
     // TODO(developer): Replace the variables before running the sample.
     // TODO: Replace with your project ID.
     String projectId = "your-project-id";
-
-    getProjectPolicy(projectId);
+    // TODO: Replace with your service account name.
+    String serviceAccount = "your-service-account";
+    getPolicy(projectId, serviceAccount);
   }
 
-  // Gets a project's policy.
-  public static Policy getProjectPolicy(String projectId) throws IOException {
+  // Gets a service account's IAM policy.
+  public static Policy getPolicy(String projectId, String serviceAccount) throws IOException {
+
+    // Construct the service account email.
+    // You can modify the ".iam.gserviceaccount.com" to match the service account name in which
+    // you want to delete the key.
+    // See, https://cloud.google.com/iam/docs/creating-managing-service-account-keys?hl=en#deleting
+
+    String serviceAccountEmail = serviceAccount + "@" + projectId + ".iam.gserviceaccount.com";
+
     // Initialize client that will be used to send requests.
     // This client only needs to be created once, and can be reused for multiple requests.
-    try (ProjectsClient projectsClient = ProjectsClient.create()) {
+    try (IAMClient iamClient = IAMClient.create()) {
       GetIamPolicyRequest request = GetIamPolicyRequest.newBuilder()
-              .setResource(ProjectName.of(projectId).toString())
+              .setResource(ServiceAccountName.of(projectId, serviceAccountEmail).toString())
               .build();
-      return projectsClient.getIamPolicy(request);
+      Policy policy = iamClient.getIamPolicy(request);
+      System.out.println("Policy retrieved: " + policy.toString());
+      return policy;
     }
   }
 }
-// [END iam_get_policy]
+// [END iam_service_account_get_policy]
